app/requirements.txt
streamlit==1.42.0
pypdf==4.0.0
pdfplumber==0.10.3
lightrag-hku==1.1.2
supabase==2.3.4
openai==1.12.0
google-generativeai==0.3.2
python-dotenv==1.0.1
tqdm==4.66.1
numpy==1.26.4
torch==2.2.1
transformers==4.38.1
datasets==2.17.0
einops==0.7.0
psycopg2-binary==2.9.9
sentence-transformers==2.5.1
nest_asyncio==1.6.0
asyncio
python-multipart
